+++ modules/client/src/tokenVoting/index.ts
@@ -1,12 +1,16 @@
 export { TokenVotingClient } from "./client";
 export {
+  DelegateTokensParams,
+  DelegateTokensStep,
+  DelegateTokensStepValue,
   Erc20TokenDetails,
   IMintTokenParams,
   ITokenVotingPluginInstall,
+  TokenVotingMember,
   TokenVotingPluginPrepareInstallationParams,
   TokenVotingProposal,
   TokenVotingProposalListItem,
   TokenVotingProposalResult,
+  UnwrapTokensStep,
   WrapTokensStep,
-  UnwrapTokensStep
 } from "./interfaces";