+++ modules/client/CHANGELOG.md
@@ -13,6 +13,11 @@ TEMPLATE:
 
 -->
 ## [UPCOMING]
+### Added
+- Add `delegate` function in `TokenVotingClient`
+### Changed
+- `getMembers` function in `TokenVotingClient` no loger returns an `string[]`, not it returns `TokenVotingMember[]` which includes the delegation status of each member
+## [1.6.0]
 ### Fixed
 - Fixed plugin version on the dao subgraph queries
 ### Added
@@ -21,9 +26,6 @@ TEMPLATE:
 
 ### Changed
 - Properties `signer` and `web3providers` are now optional in the `ContextParams`
-
-
-## [1.5.0]
 ### Added
 - Add `wrapTokens` and `unwrapTokens` functions to support token contract without support for snapshots
 - `Context.network` type changes from `Networkish` to `Network`